remove old files, add new user data stream
