+++ api/main/__init__.py
@@ -4,6 +4,7 @@ from pymongo import MongoClient
 from main.tools import JsonResp
 from jose import jwt
 import os
+from flask_socketio import send, SocketIO
 
 # Import Routes
 from main.user.routes import user_blueprint
@@ -11,6 +12,7 @@ from main.account.routes import account_blueprint
 from main.bots.routes import bot_blueprint
 from main.deals.routes import deal_blueprint
 from main.orders.routes import order_blueprint
+from main.userDataStream.routes import userDataStream_blueprint
 
 def create_app():
 
@@ -18,18 +20,10 @@ def create_app():
   app = Flask(__name__)
   app.config.from_pyfile("config/config.cfg")
   cors = CORS(app, resources={r"/*": { "origins": app.config["FRONTEND_DOMAIN"] }})
-
+  socketio = SocketIO(app)
+  socketio.run(app, debug=True)
   # Misc Config
   os.environ["TZ"] = app.config["TIMEZONE"]
-
-  # Database Config
-  # if app.config["ENVIRONMENT"] == "development":
-    # mongo = MongoClient(app.config["MONGO_HOSTNAME"], app.config["MONGO_PORT"])
-    # app.db = mongo[app.config["MONGO_APP_DATABASE"]]
-  # else:
-  #   mongo = MongoClient("localhost")
-  #   mongo[app.config["MONGO_AUTH_DATABASE"]].authenticate(app.config["MONGO_AUTH_USERNAME"], app.config["MONGO_AUTH_PASSWORD"])
-  #   app.db = mongo[app.config["MONGO_APP_DATABASE"]]
 
   mongo = MongoClient(app.config["MONGO_HOSTNAME"], app.config["MONGO_PORT"])
   mongo[app.config["MONGO_AUTH_DATABASE"]].authenticate(app.config["MONGO_AUTH_USERNAME"], app.config["MONGO_AUTH_PASSWORD"])
@@ -41,10 +35,15 @@ def create_app():
   app.register_blueprint(bot_blueprint, url_prefix="/bot")
   app.register_blueprint(deal_blueprint, url_prefix="/deal")
   app.register_blueprint(order_blueprint, url_prefix="/order")
+  app.register_blueprint(userDataStream_blueprint, url_prefix="/user-data-stream")
 
   # Index Route
   @app.route("/")
   def index():
     return JsonResp({ "status": "Online" }, 200)
   
+  @socketio.on('message')
+  def handle_message(message):
+    send(message)
+
   return app
