+++ api/main/userDataStream/user_data_stream.http
@@ -1,0 +1,3 @@
+
+## open websocket
+GET http://localhost:5000/user-data-stream/listenkey HTTP/1.1
