+++ api/run.py
@@ -1,5 +1,6 @@
 from main import create_app
 import logging
+
 app = create_app()
 
 if __name__ == "__main__":
