+++ api/test-ws.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<meta charset="utf-8" />
+<title>WebSocket Test</title>
+<script language="javascript" type="text/javascript">
+
+var wsUri = "wss://stream.binance.com:9443/api/v3/userDataStream";
+var output;
+
+function init()
+{
+  output = document.getElementById("output");
+  testWebSocket();
+}
+
+function testWebSocket()
+{
+  websocket = new WebSocket(wsUri);
+  websocket.onopen = function(evt) { onOpen(evt) };
+  websocket.onclose = function(evt) { onClose(evt) };
+  websocket.onmessage = function(evt) { onMessage(evt) };
+  websocket.onerror = function(evt) { onError(evt) };
+}
+
+function onOpen(evt)
+{
+  const subscription = {
+  "method": "SUBSCRIBE",
+  "params": [
+    "btcusdt@aggTrade"
+  ],
+  "id": 1
+}
+  writeToScreen("CONNECTED");
+  doSend(JSON.stringify(subscription));
+}
+
+function onClose(evt)
+{
+  writeToScreen("DISCONNECTED");
+}
+
+function onMessage(evt)
+{
+  console.log(evt);
+  writeToScreen('<span style="color: blue;">RESPONSE: ' + evt.data+'</span>');
+  websocket.close();
+}
+
+function onError(evt)
+{
+  writeToScreen('<span style="color: red;">ERROR:</span> ' + evt.data);
+}
+
+function doSend(message)
+{
+  writeToScreen("SENT: " + message);
+  websocket.send(message);
+}
+
+function send() {
+  const streamName = "btcusdt@aggTrade";
+  const streamUrl = `/stream?streams=${streamName}`
+  const newWs = new WebSocket(streamUrl);
+  console
+}
+
+function writeToScreen(message)
+{
+  var pre = document.createElement("p");
+  pre.style.wordWrap = "break-word";
+  pre.innerHTML = message;
+  output.appendChild(pre);
+}
+
+window.addEventListener("load", init, false);
+
+</script>
+
+<h2>WebSocket Test</h2>
+
+<div id="output"></div>
+
+<input class="draw-border" id="sendMessage" size="35" value="" />
+<button class="echo-button" id="send" onclick="send()">Send</button>
