--- app/algorithms/__init__.py
@@ -1,5 +1,0 @@
-from .sell_bb import *
-from .sell_kc import *
-from .sudden_inc import *
-from .sudden_inc_alt import *
-from .safe_than_sorry_strategy import *
