--- app/launchers/__init__.py
@@ -1,2 +1,0 @@
-from .launch_buy import *
-from .launch_sell import *
