--- app/orders/__init__.py
@@ -1,1 +1,0 @@
-from .new_order import *
