--- app/signal.log
@@ -1,2 +1,0 @@
-01:53:48,41 root INFO running sell_funds BNBBTC 
-01:53:49,266 root INFO short BNBBTC @ 0.00295790
