--- app/tests/meta_pytest.py
@@ -1,2 +1,0 @@
-def hello_test():
-    return print('hello test')
