--- app/utilities/__init__.py
@@ -1,6 +1,0 @@
-#%%
-from .environment import *
-from .candlestick import *
-from .get_data import *
-from .indicators import *
-from .api import *
