+++ requirements.txt
@@ -1,105 +1,51 @@
 appdirs==1.4.3
-appnope==0.1.0
-asn1crypto==0.24.0
-astroid==2.2.5
-attrs==19.1.0
+asn1crypto==1.2.0
+astroid==2.3.2
+attrs==19.3.0
 autopep8==1.4.4
-backcall==0.1.0
-black==19.10b0
-bleach==3.1.0
-cachetools==3.1.1
+black==19.3b0
 certifi==2019.9.11
-cffi==1.12.2
+cffi==1.13.1
 chardet==3.0.4
 Click==7.0
-conda==4.7.12
-conda-package-handling==1.6.0
-cryptography==2.6.1
-decorator==4.4.0
-defusedxml==0.6.0
-entrypoints==0.3
+cryptography==2.8
+ecdsa==0.13
 Flask==1.1.1
 Flask-Cors==3.0.8
-future==0.16.0
-google-api-python-client==1.7.11
-google-auth==1.6.3
-google-auth-httplib2==0.0.3
-google-auth-oauthlib==0.4.0
-httplib2==0.13.1
+future==0.18.2
+gevent==1.4.0
+greenlet==0.4.15
 idna==2.8
-ipykernel==5.1.2
-ipython==7.8.0
-ipython-genutils==0.2.0
-ipywidgets==7.5.1
-isort==4.3.19
+isort==4.3.21
 itsdangerous==1.1.0
-jedi==0.15.1
-Jinja2==2.10.1
-jsonschema==3.0.2
-jupyter==1.0.0
-jupyter-client==5.3.3
-jupyter-console==6.0.0
-jupyter-core==4.5.0
-lazy-object-proxy==1.4.1
+Jinja2==2.10.3
+jose==1.0.0
+lazy-object-proxy==1.4.3
 MarkupSafe==1.1.1
 mccabe==0.6.1
-mistune==0.8.4
-mkl-fft==1.0.12
-mkl-random==1.0.2
-mpld3==0.3
-nbconvert==5.6.0
-nbformat==4.4.0
-nltk==3.4.5
-notebook==6.0.1
-numpy==1.16.4
-oauthlib==3.1.0
-pandas==0.24.2
-pandocfilters==1.4.2
-parso==0.5.1
-pathspec==0.6.0
-pexpect==4.7.0
-pickleshare==0.7.5
-pipenv==2018.11.26
-prometheus-client==0.7.1
-prompt-toolkit==2.0.9
-ptyprocess==0.6.0
+mkl-fft==1.0.15
+mkl-random==1.1.0
+mkl-service==2.3.0
+numpy==1.17.3
+pandas==0.25.2
+passlib==1.7.1
 pyasn1==0.4.7
-pyasn1-modules==0.2.6
 pycodestyle==2.5.0
-pycosat==0.6.3
 pycparser==2.19
-Pygments==2.4.2
-pylint==2.3.1
+pycryptodome==3.8.2
+pylint==2.4.3
+pymongo==3.9.0
 pyOpenSSL==19.0.0
-pyrsistent==0.14.11
-PySocks==1.6.8
-python-dateutil==2.8.0
+PySocks==1.7.1
+python-dateutil==2.8.1
 python-dotenv==0.10.3
-pytz==2019.1
-pyzmq==18.1.0
-qtconsole==4.5.5
-regex==2019.11.1
-requests==2.21.0
-requests-oauthlib==1.2.0
+python-jose==3.0.1
+pytz==2019.3
+requests==2.22.0
 rsa==4.0
-ruamel-yaml==0.15.46
-schedule==0.6.0
-Send2Trash==1.5.0
 six==1.12.0
-statsd==3.3.0
-talisker==0.15.0
-terminado==0.8.2
-testpath==0.4.2
 toml==0.10.0
-tornado==6.0.3
-tqdm==4.36.1
-traitlets==4.3.2
-typed-ast==1.4.0
-uritemplate==3.0.0
-virtualenv==16.7.7
-virtualenv-clone==0.5.3
-wcwidth==0.1.7
-webencodings==0.5.1
-Werkzeug==0.15.6
-widgetsnbextension==3.5.1
-wrapt==1.11.1
+urllib3==1.24.2
+websocket-client==0.56.0
+Werkzeug==0.16.0
+wrapt==1.11.2
